requirements.txt
discord.py==2.2.3
fastapi==0.95
python-dotenv==1.0.0
uvicorn==0.22.0
pydantic~=1.10.7
aiohttp~=3.8.4
loguru==0.7.0
aiofiles==23.1.0
python-multipart==0.0.6
pandas==1.5.3
revChatGPT==6.7.6
moviepy==1.0.3
Flask==2.3.2
openai==0.27.8
Pillow==10.0.0